更改extra_repr的位置
