+++ python/paddle/nn/functional/__init__.py
@@ -70,7 +70,6 @@ from .conv import conv2d_transpose  # noqa: F401
 from .conv import conv3d  # noqa: F401
 from .conv import conv3d_transpose  # noqa: F401
 from .distance import pairwise_distance # noqa: F401
-from .distance import extra_repr # noqa: F401
 from .extension import diag_embed  # noqa: F401
 from .extension import sequence_mask
 from .loss import binary_cross_entropy  # noqa: F401
