+++ python/paddle/nn/functional/distance.py
@@ -100,14 +100,5 @@ def pairwise_distance(x, y, p=2., epsilon=1e-6, keepdim=False, name=None):
 
     return out
 
-def extra_repr(self):
-    main_str = 'p={p}'
-    if self.epsilon != 1e-6:
-        main_str += ', epsilon={epsilon}'
-    if self.keepdim != False:
-        main_str += ', keepdim={keepdim}'
-    if self.name != None:
-        main_str += ', name={name}'
-    return main_str.format(**self.__dict__)
